fix bug in gpio, cleanup, add print functions

+++ src/platform/time.h
@@ -6,6 +6,6 @@ namespace platform::time {
   uint32_t now_us();
   float now_ms();
 
-  void delay_us(uint32_t v);
-  void delay_ms(uint32_t v);
+  void delay_us(uint64_t v);
+  void delay_ms(uint64_t v);
 }
